CBP - stock Transactions - Crypto Money Restock - QA #3148 Fix Implementation BigDecimal

+++ CBP/library/api/fermat-cbp-api/src/main/java/com/bitdubai/fermat_cbp_api/layer/wallet/crypto_broker/interfaces/CryptoBrokerStockTransaction.java
@@ -2,17 +2,19 @@ package com.bitdubai.fermat_cbp_api.layer.wallet.crypto_broker.interfaces;
 
 import com.bitdubai.fermat_cbp_api.all_definition.wallet.StockTransaction;
 
+import java.math.BigDecimal;
+
 /**
  * Created by Franklin  Marcano 03.12.2015
  */
 public interface CryptoBrokerStockTransaction extends StockTransaction {
     //TODO: Documentar
 
-    float getRunningBookBalance();
+    BigDecimal getRunningBookBalance();
 
-    float getRunningAvailableBalance();
+    BigDecimal getRunningAvailableBalance();
 
-    float getPreviousBookBalance();
+    BigDecimal getPreviousBookBalance();
 
-    float getPreviousAvailableBalance();
+    BigDecimal getPreviousAvailableBalance();
 }
